Add files via upload
With Reference to 0.1.2 The the Token based authorization has been handled on every login request , on successfull authentication token is generated with current timestamp encrypted in it.
In the middleware a for the succesive request the token is validated for expiration of token with the 30 mins time interval.
In the same middleware we are sending the newly generated token with the current timestamp. This would make sure that every rest api call need not handle new token generation logic since it is handled in the middleware.